changes to way client uploads students
able to alert when a specific subject does not match.
The client can then change the name of the subject in the 'options' page.
Improved robustness so students data is not saved until all validation checks have been passed. Before, there were some left over students from failed uploads which is cumbersome to deal with.